backend/requirements.txt
praw>=7.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
googlesearch-python>=1.2.3
fastapi>=0.104.0
uvicorn>=0.24.0
googlemaps>=4.10.0
openai>=1.0.0
